feat(mappings): adds uri checks & supports non-ascii character in uri
in addition, moves testdata & adds tests
