+++ internal/mckmaps/testdata/mappings/mappings-0.json
