+++ internal/mckmaps/testdata/mappings/mappings-1.json
