+++ internal/mckmaps/testdata/mappings/mappings-2.json
