+++ internal/mckmaps/testdata/mappings/mappings-3.json
