+++ internal/mckmaps/testdata/mappings/mappings-4.json
