+++ internal/mckmaps/testdata/mappings/mappings-5.json
