+++ internal/mckmaps/testdata/mappings/mappings-6.json
