+++ internal/mckmaps/testdata/mappings/mappings-7.json
@@ -1,7 +1,7 @@
 {
   "type": "mappings",
   "mappings": {
-    "uri": "/test-for-redirects",
+    "uri": "/跳转测试",
     "policies": [
       {
         "when": {
