+++ internal/mckmaps/testdata/mappings/mappings-8.json
