+++ internal/types/number.go
@@ -1,4 +1,4 @@
-package typeutil
+package types
 
 import "math"
