+++ internal/types/number_test.go
@@ -1,4 +1,4 @@
-package typeutil
+package types
 
 import (
 	"strconv"
